requirements.txt
faiss-cpu==1.8.0
gcsfs==2024.3.1
google-auth==2.29.0
google-cloud-storage==2.16.0
numpy==1.26.4
openai==1.14.2
pandas==2.2.1
plotly==5.20.0
scikit-learn==1.4.1.post1
st-files-connection==0.1.0
streamlit==1.32.2